Move panic handler from utils to cli package
